+++ lib/core/Ioc.ts
@@ -1,5 +1,8 @@
-import { ChyanMetaKey } from "../constant/symbol";
-import { fetchMetadata, metadatas } from "./metadata-storage";
+import { trace } from "console";
+import { ChyanMetaKey } from "../constant/metakey";
+import { chyanLogger } from "../utils/chyanlog";
+import { Wires } from "./AutoWired";
+import { destory, fetchMetadata, metadatas } from "./MetadataStorage";
 
 const pool: Map<string, Injector> = new Map();
 
@@ -30,7 +33,9 @@ export async function resolve() {
 
         //查看计数器是不是已经到了最大值
         if (nowCount >= MAX_COUNT) {
-          throw "无法实例全部对象，请查看是否存在循环依赖，如果存在，请使用属性注入！";
+          // todo 尝试使用属性注入这个属性
+          chyanLogger.fatal("无法实例全部对象，请查看是否存在循环依赖，如果存在，请使用属性注入！");
+          process.exit();
         }
 
         circleMap.set(node.id, nowCount + 1);
@@ -46,32 +51,54 @@ export async function resolve() {
       }
     }
 
+    const prepareHandler = fetchMetadata<Function>(ChyanMetaKey.prepare, node.target);
+
+    const packingHandler = fetchMetadata<Function>(ChyanMetaKey.packing, node.target);
+
+    const doneHandler = fetchMetadata<Function>(ChyanMetaKey.done, node.target);
+
     const handleBeforeIns = fetchMetadata<Function>(ChyanMetaKey.beforeIns, node.target);
 
-    if (handleBeforeIns !== null) {
-      await handleBeforeIns(node.target, args);
-    }
+    if (handleBeforeIns !== null) await handleBeforeIns(node.target, args);
 
     const handleOnIns = fetchMetadata<Function>(ChyanMetaKey.onIns, node.target);
 
     let changedObj: object | null = null;
-    if (handleOnIns !== null) {
-      changedObj = await handleOnIns(node.target, args);
-    }
+
+    if (handleOnIns !== null) changedObj = await handleOnIns(node.target, args);
 
     const o = changedObj ?? Reflect.construct(node.target, args);
 
     const handleAfterIns = fetchMetadata<Function>(ChyanMetaKey.afterIns, node.target);
 
-    if (handleAfterIns !== null) {
-      await handleAfterIns(o);
-    }
+    if (handleAfterIns !== null) await handleAfterIns(o);
 
-    pool.set(node.id, { id: node.id, instance: o });
+    const wires = fetchMetadata<Wires[]>(ChyanMetaKey.wires, node.target) ?? [];
+
+    const notPushPool = fetchMetadata<boolean>(ChyanMetaKey.noninvasive, node.target);
+
+    notPushPool || pool.set(node.id, { id: node.id, instance: o, wires });
   }
+
+  for (const [id, injector] of pool) {
+    for (const wire of injector.wires) {
+      const id = fetchMetadata<string>(ChyanMetaKey.id, wire.type);
+      if (id) {
+        injector.instance[wire.key] = pool.get(id)?.instance;
+      } else {
+        chyanLogger.fatal(`${injector.instance.constructor}的属性注入${wire.type.name}不存在！`);
+        process.exit();
+      }
+    }
+  }
+}
+
+export function fetchInjector<T>(id: string): T | null {
+  return ((pool.get(id)?.instance as unknown) as T) || null;
 }
 
 interface Injector {
   id: string;
   instance: object;
+  wires: Wires[];
 }