+++ lib/decorators/Routes/metakey.ts
@@ -1,0 +1,9 @@
+export enum RouteMetaKey {
+  middlewares = "chyan:router-middlewares",
+  prefix = "chyan:router-prefix",
+  suffix = "chyan:router-suffix",
+  method = "chyan:router-method",
+  actions = "chyan:router-actions",
+  args = "chyan:router-args",
+  pipe = "chyan:router-pipe",
+}