+++ lib/decorators/Routes/pipe.ts
@@ -2,8 +2,8 @@ import { attachMetadata } from "../../core/MetadataStorage";
 import { Constructor } from "../../types/types";
 import { RouteMetaKey } from "./metakey";
 
-export function Middlewares(middlewares: Function[]) {
+export function Pipe(pipe: Function) {
   return (target: Constructor | object, key?: string, desc?: PropertyDescriptor) => {
-    attachMetadata(RouteMetaKey.middlewares, middlewares, target, key);
+    attachMetadata(RouteMetaKey.pipe, pipe, target, key);
   };
 }