+++ lib/decorators/Routes/share.ts
@@ -1,8 +1,0 @@
-export enum RouteMetaKey {
-  middlewares = "chyan:middlewares",
-  prefix = "chyan:prefix",
-  suffix = "chyan:suffix",
-  method = "chyan:method",
-  actions = "chyan:actions",
-  args = "chyan:args",
-}