+++ lib/types/types.ts
@@ -1,1 +1,9 @@
-export type Constructor = new (...args: any[]) => object;
+import { Context, Next } from "koa";
+
+export type Constructor<T = object> = new (...args: any[]) => T;
+
+export type ChyanContext = Context;
+
+export type ChyanNext = Next;
+
+export { Document, Schema, SchemaOptions } from "mongoose";