+++ lib/utils/CheckClassType.ts
@@ -1,0 +1,4 @@
+export function isClass(fn: any) {
+  const str = fn.toString();
+  return str.slice(0, 5) === "class";
+}